feat(ui): add plant image and water status/button
Modified the UI of the garden screen to make it
display the actual image of a plant.
Added a "water bar" that will display the current
water status of a plant as well as a "water
button" that users can press to notify the app
that they have watered this plant.