Feature/update garden backend (#94)
* feat: make watering bar proportional in status

Made the watering bar proportional to the current
value in status. Modified the health calculator to
make it keep track of the current percentage in
which the plant is inside its current status.
Also modified the watering button to make it
actually water the plant !

* test: add tests for float generation for WaterBar

Improved the tests by adding new tests for the
float generation part just added.
Also improved the code with minor changes.

* chore(nav): navigate when clicking on plant card

Added the possibility for the user to go on the
EditPlant screen when pressing on a plant card in
the user's garden. Modified accordingly the
Screen, AppNavHost and MainActivity in
consequence.

* fix(test): fix garden screen UI tests

Made garden screen tests use the useUnmergedTree
parameter as the plant cards are now clickable.

* chore(format): apply ktfmt to the modified  tests

* chore: addressed PR review comments

* test(nav): add garden/editPlant navigation tests

Added new tests for navigation between the garden
screen and the edit plant screen.

* feat: make user row actually depend on user info

Added the profile repository to the garden screen
view model and made the actual name and avatar
displayed on the screen depend on the user ones
from the repository. Also added tests for this new
behaviour and adapted the previous ones to fit
with the new implementation.

* chore: addressed PR review comments