Feature/plant deletion confirmation (#105)
* feat(ui): add plant deletion confirmation popup

Added a new popup that displays when users
request the deletion of their plant in the edit
plant screen.
Note: Tests are not already included.

* test: add new popup ui and navigation tests

Added several tests that test the popup displays
correctly when requested as well as some tests
for the navigation between the EditPlant, Garden
and DeletePlantPopup.

* chore: addressed PR review comments

* test: fix test with newly fetched features

Fixed tests on the main branch that does not take
into account the deletion popup.