Feature/push notifications (#144)
* feat: add basic notifications permission handling

Handled the notifications permission for API 33 or
greater. This simple handling assume that the user
will accept the asked permission. The complete
workflow will be added later on.
Also added Firebase Messaging Service in the class
PushNotificationsService (empty for now).

* feat: add firebase function for plant updating

Configured Firebase Cloud Functions for the app's
Firebase project and added a first cloud function
that refreshes all database plants every hours.
Also created a firestore.rules file that contains
current Firestore rules on the Firebase project.

* feat(FCM): add FCM new token handling

Handled new token generation via onNewToken of the
FirebaseMessagingService implementation. The app
tries to upload the new token on the user's remote
profile if possible, otherwise stores it locally
temporarly and tries again later. Added a new
FCMTokenSyncer object to handle token
synchronization between the locally stored token
and the Firestore repository.
The behavious has been working but is currently
not tested by unit or android tests.

# Conflicts:
#	app/src/main/java/com/android/mygarden/ui/profile/ProfileScreenComposables.kt
#	app/src/main/java/com/android/mygarden/ui/profile/ProfileViewModel.kt

* feat: add push notifications automation

Added new Cloud Functions. The primary goal of
all these new functions is the send push
notifications to users when their plants
transition to the health status "NEED_WATER" or
"CRITICALLY_DRY". Tests will come accordingly.

* feat: handle notifications app side

Added notifications handling on the app side. The
notifications are shown only if users are not
on the app currently (otherwise a popup is shown)
and when clicking on one notification, users are
redirected directly on the garden screen.

* bugfix: modify notification type sent

Fixed the type of the notification that is sent.
Before was AndroidNotification but now is a plain
notification and does not need to carry on
"click_action" field to go on the garden screen.

* test: add cloud functions testing

Added cloud functions testing for next_status
computing and all plant status updating
(including) FCM messages sending. The whole
Firestore database and FCM send function are
mocked.

* chore(git): modify gitignore

Made .gitignore ignore virtual environements
and pycaches.

* chore: add token fun in profile repo interface

Added definition of attachFCMToken and
getFCMToken in the interface rather than directly
in the Firestore one.

* fix: make profile repo implementations work

Add basic implementations of the new functions
added for the ProfileRepository.

* test: add tests for attachFCMToken and getFCMToken

* test: add tests for onNewToken()

* chore(format): apply ktfmt to the modified code

* fix: fix missing test

Fixed the commented test to work fine.

* fix: M1end2end permissions and firestore rules

Give notifications permission in M1 end2end.
Update more laxist rules.

* fix: ensure test environnement for intent in main

Fixed end2end tests and popupBackendTests by
handling a test environnement in the MainActivity.

* fix: replace hard-coded strings notifications

Added elements in the PushNotificationsService
companion object to include strings used for
better code.

* chore: addressed PR review comments and add test

* chore(format): apply ktfmt to the modified code

* fix(manifest): try to add flutter again

---------

Co-authored-by: Simonperraudin <saimp02@gmail.com>