Feature/edit plant update (#162)
* feat: add location editing on EditPlantScreen

Added a new field on the EditPlantScreen where
users can chose between the 3 locations (INDOOR,
OUTDOOR, UNKNOWN) in a drop down menu if the plant
has not been recognized by the API.

# Conflicts:
#	app/src/main/java/com/android/mygarden/ui/editPlant/EditPlantScreen.kt

* test: add tests for the new implementation

New tests include compose test for the UI
components as well as a unit test for the
EditPlantViewModel.

* feat: add a text field for the light exposure

The same work has been done for the light exposure
field. Only modifiable when the plant is not
recognized.

* test: adapt and add compose test for lightExposure

Adapted the compose tests to handle the new fields
and added tests for the lightExposure.

* chore: apply formatting and add viewModel tests

* chore: address PR review changes

* chore: adapt code after merging main

* test(ui): try to fix UI tests

* test(end2end): fix end2end tests