Feature/plant creation date (#165)
* feat: add a dateOfCreation field in ownedPlant

Added this field to contains the information about
when the plant was added to users' gardens.
Handled serialization/deserialization in the
FirestoreMapper object. Added some simple unit
tests to ensure the behaviour works.

* feat(ui): add UI representation of creation date

Added a new small field in the newly created
PlantInfoScreen that is between the garden and
the EditPlantScreen (ownedPlant not null).
Applied format to modified files on the branch.

* test(ui): add test for the added UI components

Add simple compose tests for the new UI component
in the newly updated PlantInfoScreen. Also managed
the hard-coded strings in the UI.

* test: adapt existing failing test

* chore: address PR review feedback