fix(ui): improve settings access UX
Fixed the UI as it was not UX friendly as stated
in the M2 feedback. Changed the plain text with a
button and better text with a icon. Also replaced
the hard-coded color at the top of the screen by
using the Extended theme that was added few weeks
ago.