Feature/notifications permission workflow (#180)
* chore: rename misnamed function

Renamed the function
"setHasAlreadyDeniedCameraPermission()" with a
capital to respect best naming conventions.

* feat: add notifications permission workflow

Created a new object that handles the
notifications permission (to avoid polluting the
MainActivity with backend logic). It implements
the notifications permission workflow that only
asks for the notifications permission once and
let the user later change in the settings.

* test: add unit tests for the permission workflow

Added some basic unit tests that ensure that the
workflow works. As the testing part has been
implemented such that no intent is launched when
executing the MainActivity in a test environment,
the verification that the intent actually launches
has been made manually.

* test: add tests that check permission request

Added tests that actually tests the launch of the
permission request via the launcher. Made possible
by changing the AskForNotificationsPermission()
modifier from private to internal.

* chore: address PR review feedback