UI/upgrade plant info screen (#189)
* ui: enhance PlantInfoScreen UI

Added a new version of the PlantInfoScreen UI that
is more UX friendly. Added new cards, icons,
better text formatting, etc.
Also added a space between the Icon and the Text
in the SignInScreen.

* feat(ui): add responsive location Icon

Added a responsive location Icon for the different
locations in a separated function. Also applied
ktfmt to format the code.

* chore: remove hard-coded strings

Removed the hard-coded strings present in the UI
with string in ui_strings.xml for good practice.

* test(ui): adapt existing tests and add new ones

Adapt the current tests so that they work with the
new UI and added some new ones for the new texts.

# Conflicts:
#	app/src/androidTest/java/com/android/mygarden/zEndToEnd/EndToEndM1.kt

* chore(ui): add constants UI paddings, sizes, etc.

* chore: address PR review feedback