feat: add achievements definition and repository
Created new types for the achievements handling.
Added a generic achievement type as well as
specific types for the first achievements we
decided to add.
Moreover, added a new repository that handles the
achievement in Firestore at the user level (i.e
each user has a collection user/achievements).
Added getter, setters as well as a flow that will
later be retrieved in the ViewModel of the
AchievementsScreen.
Also added Firestore tests for the repository and
unit tests for the definitions.