Feature/achievement backend (#211)
* feat: add achievements definition and repository

Created new types for the achievements handling.
Added a generic achievement type as well as
specific types for the first achievements we
decided to add.
Moreover, added a new repository that handles the
achievement in Firestore at the user level (i.e
each user has a collection user/achievements).
Added getter, setters as well as a flow that will
later be retrieved in the ViewModel of the
AchievementsScreen.
Also added Firestore tests for the repository and
unit tests for the definitions.

* feat: implement the first 3 types of achievements

Added 3 new types of achievements: plants number,
friends number and healthy streak.
Created the logic for these new achievements type
that update the newly created achievements repo.

# Conflicts:
#	app/src/main/java/com/android/mygarden/model/friends/FriendsRepositoryFirestore.kt

# Conflicts:
#	app/src/main/java/com/android/mygarden/ui/profile/ProfileScreenComposables.kt

# Conflicts:
#	app/src/test/java/com/android/mygarden/ui/feed/FeedScreenTests.kt

* feat(ui): create a first AchievementsScreen

Created a first AchievementsScreen and
AchievementsViewModel to manually test the
emission of flow. The AchievementsViewModel is
almost fully implemented but the screen is only a
mockup and is not integrated in the app for now.
The UI and navigation will be done next week.

* feat: update Firebase functions

Update Firebase function to work accordingly with
the healthy streak achievement. Update the
healthySince field at each scheduled status
update. Added simple python test for the function.

* chore: adapt achievement activity

Adapted the ActivityAchievement to contain a
AchievementType instead of a plain String name.
Updated tests and depoendencies accordingly.

* fix(test): fix unit tests that use provider

Created a FakeAchievementsRepository to handle
unit tests that needs to have a local test repo
for the achievements (no Firestore interaction).

* fix(test): make updatePlantStatus same as in repo

Changed the private function in tests to be the
same as the one in the repository (that has been
modified for the achievements implementation).

* fix(test): fix previous plant rep test

Fixed outdated plant repository tests to ignore
the healthySince field. Added also serialization
tests for the new healthySince field.

* fix(test): fix editOwnedPlant_works test

* fix(test): fix navigation tests

Fixed navigation tests by adding a
FakeAchievementsRepository in the androidTest.
Also added the @Keep annotation on the relevant
enum and data classes for the AKP build.

* chore: apply ktfmt format

* fix(ui): fix Activities to use achievement types

* test: add AchievementsViewModel tests

Added small AchievementsViewModel tests. Also
applied the ktfmt to the previous code.

* chore: address PR review feedback