feat(ui): implement AchievementsScreen
Implemented the AchievementsScreen and adapted the
corresponding AchievementsViewModel. The screen
now displays each type of achievements with a
level progression on a card. Each card is
clickable and when clicking on it the user has
more info about the achievement and their
progression for it (what to do for next level).