feat(ui): fusion Garden and Achievements screens
Created a new ParentTabScreenGarden that displays
either the GardenScreen either the
AchievementsScreen depending on the selected tab.
Moved the Scaffold in the parent screen, which now
displays the profile name, avatar, log out button
and a newly created tab that allows the user to
switch between achievements and garden easily.
In the top bar of the garden the edit profile
button has been removed, and the user now needs to
click on the avatar or name to edit the profile.
Also updated the navigation and the possibility to
see friend's garden, which now also allow users to
see their friend's achievements.