Feature/achievements screen (#236)
* feat(ui): implement AchievementsScreen

Implemented the AchievementsScreen and adapted the
corresponding AchievementsViewModel. The screen
now displays each type of achievements with a
level progression on a card. Each card is
clickable and when clicking on it the user has
more info about the achievement and their
progression for it (what to do for next level).

* chore: improve ViewModel and add documentation

Improved the ViewModel by removing repetition
code. Added proper documentation both the Screen
and the ViewModel.

* feat(ui): fusion Garden and Achievements screens

Created a new ParentTabScreenGarden that displays
either the GardenScreen either the
AchievementsScreen depending on the selected tab.
Moved the Scaffold in the parent screen, which now
displays the profile name, avatar, log out button
and a newly created tab that allows the user to
switch between achievements and garden easily.
In the top bar of the garden the edit profile
button has been removed, and the user now needs to
click on the avatar or name to edit the profile.
Also updated the navigation and the possibility to
see friend's garden, which now also allow users to
see their friend's achievements.

* chore(ui): document ParentScreen

Documented the ParentTabScreenGarden with full
KDoc. Moves the ProfileRow composable in the
parent screen as it is shared between the two tabs
of the parent screen.

* fix: update friends achievement

Updated the FriendsRepositoryFirestore to support
the new way of adding friends (request). Also
applied ktfmt to the code.

* chore(ui): add constants for padding, sizing, etc.

* test: fix existing tests to include new UI layout

Updated all navigation tests, and modified garden
tests to include the new parent screen.

* test: fix remaining garden test

* test(ui): add UI and navigation tests

Added full compose UI AchievementsScreenTests for
the AchievementsScreen. Also added full navigation
tests for the parent screen. Adapted the remaining
tests.

* test: add non-commited tests

* chore: remove unused value

* test: add viewModel tests

* chore: fix sonar issues