fix/healthy-streak-logic (#256)
* fix: try to fix heath status computation

This is clearly not the last version, no test have
been changed and the Firebase function is still
the same. This is just a first draft.

* test: add testing for new implementation

Added full testing of the new implementation with
short, middle and long watering frequecies and
various screnarios of overwatering.
Also adapted the overwatered UI float to be
computed the right way and tested.

* fix: update Firebase functions accordingly

After updating the logic and the tests for the
Kotlin function, updated the Firebase functions
accordingly to behave the same as the Kotlin one.
Modified and added unit tests to ensure that the
behaviour is the same.

* fix(test): fix previous tests for new behaviour

* chore: address PR review and fix previous tests