Feature/complete offline mode (#270)
* feat: make friend's garden not accesible offline

Made not possible to access a friend's garden
when in offline mode. Also removed the possibility
to delete a friend while offline for consistency.

* test: add tests for the offline mode

Added tests for the offline mode of the
FriendListScreen. Now, when offline the user can
no longer click on a friend's profile to access
their garden and no longer delete a friend.
Those cases are checked with the tests and they
also ensure the right Toast is displayed.