Feature/achievements in feed (#271)
* chore: add a new field in the GardenActivity

Added a new field levelReached to know which level
has been reached by the user when a activity is
displayed. Adapted serialization and tests.

* feat: handle achievement activities via Firebase

Handled the fact that a user getting a achievement
should be displayed in the feed screem. This
has been done via Firebase Functions. A new
function "observe" the achievements Firebase
collection and create a new activity document if
the change has made the user gain a level.
Also separated the big previous main.py into many
smaller files (as suggested during feedback) for
better code quality.

* test: add tests for new Firebase function

Added new python tests for the newly added
functions in the file achievement_activities.py.
Also fixed the timestamp that was a Firestore
timestamp instead of a epoch (long/int).

* feat(ui): adapt achievements card with level

Adapted the text displayed in the achievement card
to display the level that has been reached.

* chore: address PR review feedback